+++ projects/js-packages/licensing/components/activation-screen-success-info/product-link/index.jsx
@@ -31,10 +31,6 @@ const ProductLink = props => {
 			text: __( 'View latest backup', 'jetpack' ),
 			redirectSource: 'jetpack-license-activation-success-backup',
 		},
-		jetpack_starter: {
-			text: __( 'View latest backup', 'jetpack' ),
-			redirectSource: 'jetpack-license-activation-success-backup',
-		},
 		default: null,
 	};