+++ projects/packages/sync/src/class-package-version.php
@@ -12,7 +12,7 @@ namespace Automattic\Jetpack\Sync;
  */
 class Package_Version {
 
-	const PACKAGE_VERSION = '1.50.2-alpha';
+	const PACKAGE_VERSION = '1.50.2';
 
 	const PACKAGE_SLUG = 'sync';