+++ projects/plugins/boost/app/assets/src/js/pages/settings/sections/Modules.svelte
@@ -170,24 +170,8 @@
 		</p>
 	</Module>
 
-	<div class="settings">
-		<Module slug="image_guide">
-			<h3 slot="title">{__( 'Image Guide', 'jetpack-boost' )}<span class="beta">Beta</span></h3>
-			<p slot="description">
-				{__(
-					`This feature helps you discover images that are too large. When you browse your site, the image guide will show you an overlay with information about each image's size.`,
-					'jetpack-boost'
-				)}
-			</p>
-			<!-- svelte-ignore missing-declaration -->
-			{#if false === Jetpack_Boost.site.canResizeImages}
-				<ResizingUnavailable />
-			{/if}
-		</Module>
-	</div>
-
 	<Module slug="minify_js">
-		<h3 slot="title">{__( 'Concatenate JS', 'jetpack-boost' )}<span class="beta">Beta</span></h3>
+		<h3 slot="title">{__( 'Concatenate JS', 'jetpack-boost' )}</h3>
 		<p slot="description">
 			{__(
 				'Scripts are grouped by their original placement, concatenated and minified to reduce site loading time and reduce the number of requests.',
@@ -207,7 +191,7 @@
 	</Module>
 
 	<Module slug="minify_css">
-		<h3 slot="title">{__( 'Concatenate CSS', 'jetpack-boost' )}<span class="beta">Beta</span></h3>
+		<h3 slot="title">{__( 'Concatenate CSS', 'jetpack-boost' )}</h3>
 		<p slot="description">
 			{__(
 				'Styles are grouped by their original placement, concatenated and minified to reduce site loading time and reduce the number of requests.',
@@ -226,24 +210,8 @@
 		</div>
 	</Module>
 
-	<Module slug="image_size_analysis">
-		<h3 slot="title">
-			{__( 'Image Size Analysis', 'jetpack-boost' )}<span class="beta">Beta</span>
-		</h3>
-		<p slot="description">
-			{__(
-				`This tool will search your site for images that are too large and have an impact your visitors experience, page loading times, and search rankings. Once finished, it will give you a report of all improperly sized images with suggestions on how to fix them.`,
-				'jetpack-boost'
-			)}
-		</p>
-
-		<svelte:fragment slot="meta">
-			<RecommendationsMeta />
-		</svelte:fragment>
-	</Module>
-
 	<Module slug="image_cdn">
-		<h3 slot="title">{__( 'Image CDN', 'jetpack-boost' )}<span class="beta">Beta</span></h3>
+		<h3 slot="title">{__( 'Image CDN', 'jetpack-boost' )}</h3>
 		<p slot="description">
 			{__(
 				`Deliver images from Jetpack's Content Delivery Network. Automatically resizes your images to an appropriate size, converts them to modern efficient formats like WebP, and serves them from a worldwide network of servers.`,
@@ -251,6 +219,38 @@
 			)}
 		</p>
 	</Module>
+
+	<div class="settings">
+		<Module slug="image_guide">
+			<h3 slot="title">{__( 'Image Guide', 'jetpack-boost' )}<span class="beta">Beta</span></h3>
+			<p slot="description">
+				{__(
+					`This feature helps you discover images that are too large. When you browse your site, the image guide will show you an overlay with information about each image's size.`,
+					'jetpack-boost'
+				)}
+			</p>
+			<!-- svelte-ignore missing-declaration -->
+			{#if false === Jetpack_Boost.site.canResizeImages}
+				<ResizingUnavailable />
+			{/if}
+		</Module>
+
+		<Module slug="image_size_analysis" toggle={false}>
+			<h3 slot="title">
+				{__( 'Image Size Analysis', 'jetpack-boost' )}<span class="beta">Beta</span>
+			</h3>
+			<p slot="description">
+				{__(
+					`This tool will search your site for images that are too large and have an impact your visitors experience, page loading times, and search rankings. Once finished, it will give you a report of all improperly sized images with suggestions on how to fix them.`,
+					'jetpack-boost'
+				)}
+			</p>
+
+			<svelte:fragment slot="meta">
+				<RecommendationsMeta />
+			</svelte:fragment>
+		</Module>
+	</div>
 
 	<SuperCacheInfo />
 </div>
@@ -261,8 +261,8 @@
 		padding-top: 20px;
 	}
 	.beta {
-		background: hsl( 0, 0%, 90% );
-		color: hsl( 0, 0%, 20% );
+		background: var( --jp-green-5 );
+		color: var( --jp-green-60 );
 		padding: 2px 5px;
 		border-radius: 3px;
 		font-size: 0.8rem;