+++ projects/js-packages/partner-coupon/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.51 - 2023-07-05
+### Changed
+- Updated package dependencies. [#31659]
+
 ## 0.2.50 - 2023-06-26
 ### Changed
 - Updated package dependencies. [#31468]