+++ projects/packages/stats-admin/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.12.0 - 2023-07-05
+### Added
+- Stats Admin: add plan and product for site [#31663]
+
 ## 0.11.0 - 2023-06-26
 ### Added
 - Stats Admin: add subscribers endpoints. [#31394]