+++ projects/packages/tracking/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.13] - 2023-07-06
+### Added
+- Add Jetpack Autoloader package suggestion. [#29988]
+
 ## [1.14.12] - 2023-01-11
 ### Changed
 - Updated package dependencies.
@@ -258,6 +262,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Create package for Jetpack Tracking
 
+[1.14.13]: https://github.com/Automattic/jetpack-tracking/compare/v1.14.12...v1.14.13
 [1.14.12]: https://github.com/Automattic/jetpack-tracking/compare/v1.14.11...v1.14.12
 [1.14.11]: https://github.com/Automattic/jetpack-tracking/compare/v1.14.10...v1.14.11
 [1.14.10]: https://github.com/Automattic/jetpack-tracking/compare/v1.14.9...v1.14.10